ecommerce-api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pymongo==4.5.0
motor==3.3.1
pydantic>=2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0.post2
python-dotenv==1.0.0